pyproject.toml
[project]
name = "get-photos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-playwright>=0.7.0",
    "requests>=2.32.4",
]
